requirements.txt
asgiref==3.7.2
boto3==1.34.144
cachetools==5.4.0
celery==5.4.0
django==5.0.8
django-cache-memoize==0.2.0
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-db-geventpool==4.0.6
django-debug-toolbar
django-debug-toolbar-force
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-imagekit==5.0.0
django-model-utils==4.5.1
django-redis==5.4.0
django-s3-storage==0.15.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
docutils==0.21.2
drf-yasg[validation]==1.21.7
firebase-admin==6.5.0
flower==2.0.1
gunicorn[gevent]==22.0.0
hexbytes==0.3.1
hiredis==2.3.2
packaging>=21.0
pika==1.3.2
pillow==10.4.0
psycogreen==1.0.2
psycopg2==2.9.9
redis==5.0.7
requests==2.32.3
safe-eth-py[django]==6.0.0b37
web3==6.20.2
